requirements.txt
django
djangorestframework-simplejwt
djangorestframework
cloudinary
psycopg2-binary
python-dotenv
django-cors-headers
weasyprint
reportlab
resend
django-extensions